Move LongLongHashMap and LongPriorityQueue to netty-common
