+++ common/src/main/java/io/netty/util/internal/LongPriorityQueue.java
@@ -13,15 +13,15 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.buffer.api.pool;
+package io.netty.util.internal;
 
 import java.util.Arrays;
 
 /**
- * Internal primitive priority queue, used by {@link PoolChunk}.
+ * Internal primitive priority queue, used by {@code PoolChunk}.
  * The implementation is based on the binary heap, as described in Algorithms by Sedgewick and Wayne.
  */
-final class LongPriorityQueue {
+public final class LongPriorityQueue {
     public static final int NO_VALUE = -1;
     private long[] array = new long[9];
     private int size;
